On teleporting, players must have the correct amount of an item/xp/money
as specified in the config, with the amount being overwritable by the
attribute service. This does one check when the player triggers the
teleport, and another right before they get teleported just in case they
no longer have enough in between that time to try to exploit it.

+++ src/main/kotlin/dev/mizarc/waystonewarps/application/services/PlayerAttributeService.kt
@@ -23,7 +23,7 @@ interface PlayerAttributeService {
      * @param playerId The target player's id.
      * @returns The cost amount.
      */
-    fun getTeleportCost(playerId: UUID): Int
+    fun getTeleportCost(playerId: UUID): Double
 
     /**
      * Gets how long it takes for the player to teleport.
